# frontend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "api-server",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/{slug}/authors/{author_id}": {
      "get": {
        "tags": [
          "Authors"
        ],
        "operationId": "routes_authors_get_author",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorWithStats"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorWithStats": {
        "type": "object",
        "required": [
          "email",
          "email_count",
          "id",
          "mailing_lists",
          "name_variations",
          "thread_count"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "canonical_name": {
            "type": "string",
            "nullable": true
          },
          "first_seen": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "email_count": {
            "type": "integer",
            "format": "int64"
          },
          "thread_count": {
            "type": "integer",
            "format": "int64"
          },
          "first_email_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "last_email_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "mailing_lists": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name_variations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}